fix bugs

+++ src/js/command-pipeline/commands/CommandLineAdd.ts
@@ -16,11 +16,14 @@ class CommandLineAdd extends BaseCommand {
     const { x, y } = context.getTextCursorPosition()
     const textLength = context.getLineLength(y)
 
-    this._content = context.getContentInSelections([{ rangeX: new Range(x, textLength), rangeY: new Range(y, y) }])
-
-    context.deleteTextInRange(y, new Range(x, textLength))
     context.addNewLinesInRange(new Range(y + 1, y + 1 + this._count))
-    context.addContent({ x: 0, y: y + this._count }, this._content)
+
+    if (x < textLength) {
+      this._content = context.getContentInSelections([{ rangeX: new Range(x, textLength), rangeY: new Range(y, y) }])
+      context.deleteTextInRange(y, new Range(x, textLength))
+      context.addContent({ x: 0, y: y + this._count }, this._content)
+    }
+
     context.setTextCursorPosition({ x: 0, y: y + this._count })
     context.updateTextRepresentation()
     context.updateTextCursorPosition()
@@ -31,7 +34,11 @@ class CommandLineAdd extends BaseCommand {
     const textLength = context.getLineLength(textCursorY)
 
     context.deleteLinesInRange(new Range(textCursorY - this._count, textCursorY))
-    context.addContent({ x: textLength, y: textCursorY - this._count }, this._content)
+
+    if (this._content !== null) {
+      context.addContent({ x: textLength, y: textCursorY - this._count }, this._content)
+    }
+
     context.setTextCursorPosition({ x: textLength, y: textCursorY - this._count })
     context.updateTextRepresentation()
     context.updateTextCursorPosition()
